+++ rustracer-core/src/math/sphere.rs
@@ -1,4 +1,4 @@
-use std::f64;
+use std::f32;
 use std::cmp::PartialEq;
 use std::fmt;
 use serde::{Deserialize, Serialize};
@@ -7,19 +7,19 @@ use super::vector::Vector;
 #[derive(Copy, Clone, Deserialize, Serialize, Debug)]
 pub struct Sphere {
     pub center: Vector,
-    pub radius: f64,
+    pub radius: f32,
     pub material_index: usize,
 }
 
 impl Sphere {
-    pub fn new(center: Vector, radius: f64, material_index: usize) -> Self {
+    pub fn new(center: Vector, radius: f32, material_index: usize) -> Self {
         Sphere { center: center, radius: radius, material_index: material_index }
     }
 }
 
 impl PartialEq for Sphere {
     fn eq(&self, other: &Sphere) -> bool {
-        self.center == other.center && f64::abs(self.radius - other.radius) < f64::EPSILON
+        self.center == other.center && f32::abs(self.radius - other.radius) < f32::EPSILON
     }
 }
